requirements.txt
asgiref==3.7.2
black==23.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.4
cryptography==41.0.1
defusedxml==0.7.1
Deprecated==1.2.14
Django==4.2.3
django-github-oauth==2021.6.5
django-oauth-toolkit==2.3.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
ecdsa==0.18.0
idna==3.4
inflection==0.5.1
jwcrypto==1.5.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.1
psycopg2==2.9.6
pyasn1==0.5.0
pycparser==2.21
PyJWT==2.7.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.0.3
wrapt==1.15.0
